Moved Ghosts and Pacman into Model directory
