+++ Pacman.Tests/WorldBuilderTests.cs
@@ -1,10 +1,10 @@
 using System.Linq;
 using Moq;
 using Pacman.Business.Control;
-using Pacman.Business.Control.Ghosts;
 using Pacman.Business.Control.Selector;
 using Pacman.Business.Control.WorldLoader;
 using Pacman.Business.Model;
+using Pacman.Business.Model.Ghosts;
 using Xunit;
 
 namespace Pacman.Tests;
