+++ Pacman/Business/Control/WorldBuilder.cs
@@ -1,6 +1,6 @@
-using Pacman.Business.Control.Ghosts;
 using Pacman.Business.Control.WorldLoader;
 using Pacman.Business.Model;
+using Pacman.Business.Model.Ghosts;
 using Pacman.Variables;
 
 namespace Pacman.Business.Control;
