+++ Pacman/Business/Model/Ghosts/GhostFactory.cs
@@ -1,9 +1,8 @@
 using Pacman.Business.Control.MoveStrategies;
 using Pacman.Business.Control.Selector;
-using Pacman.Business.Model;
 using Pacman.Variables;
 
-namespace Pacman.Business.Control.Ghosts;
+namespace Pacman.Business.Model.Ghosts;
 
 public class GhostFactory
 {
